[FIX] wingpanel-more: hack to fix bad curving
It seems there have been some changes in the CSS
for Freya which removed some spacing before and
after the first and last indicator respectively.

Because of this, the panel width was slimmer,
which resulted in the curving masks cutting in
way too close to the indicators. This was actually
the reason why I created the fourth curved option
in the first place because all the others looked
broken and I couldn't figure out why.

As I don't want people to resort to CSS theme
edition in order to fix this bug, I fixed it
in the code directly.

I might rework the last curving option then, as
the visual result is now slightly different than
what is was originaly. I'm not yet sure if it
looked better without the fix or not.